refactor: エントリーポイントの削除
