+++ src/components/common_part/MarkdownForm.tsx
@@ -1,20 +1,12 @@
 import React, { useRef } from "react";
-import {
-  Tabs,
-  TabList,
-  Tab,
-  TabPanels,
-  TabPanel,
-  Stack,
-  HStack,
-  Button,
-  Icon,
-} from "@chakra-ui/react";
+import { Tabs, TabList, Tab, TabPanels, TabPanel, Stack, HStack, Button, Icon } from "@chakra-ui/react";
 import { BsImage } from "react-icons/bs";
 import { AiFillGithub } from "react-icons/ai";
 import { getStorage, ref, getDownloadURL } from "firebase/storage";
 
-import { MarkdownInput, MarkdownPreview, postImage } from "../index";
+import MarkdownInput from "./MarkdownInput";
+import MarkdownPreview from "./MarkdownPreview";
+import PostImage from "./PostImage";
 import app from "../../base";
 
 const storage = getStorage(app);
@@ -47,9 +39,7 @@ const MarkdownForm = (props: MarkdownFormProps): JSX.Element => {
   };
 
   // 画像が選択されたらマークダウンの末尾に画像を表示するマークダウンを追加
-  const handleImage: React.ChangeEventHandler<HTMLInputElement> = async (
-    event
-  ) => {
+  const handleImage: React.ChangeEventHandler<HTMLInputElement> = async (event) => {
     // ファイルが選択されているかチェック
     if (event.target.files != null && event.target.files[0] != null) {
       // FileをStorageに保存する
@@ -67,46 +57,25 @@ const MarkdownForm = (props: MarkdownFormProps): JSX.Element => {
   return (
     <Stack pb="8">
       <Tabs variant="unstyled">
-        <Stack
-          flexDir={{ base: "column-reverse", sm: "row" }}
-          justify="space-between"
-        >
+        <Stack flexDir={{ base: "column-reverse", sm: "row" }} justify="space-between">
           <TabList pt="2">
-            <Tab
-              rounded="full"
-              fontSize={{ base: "sm", md: "md" }}
-              _selected={{ color: "#FCFCFC", bg: "#99CED4" }}
-            >
+            <Tab rounded="full" fontSize={{ base: "sm", md: "md" }} _selected={{ color: "#FCFCFC", bg: "#99CED4" }}>
               マークダウン
             </Tab>
-            <Tab
-              rounded="full"
-              fontSize={{ base: "sm", md: "md" }}
-              _selected={{ color: "#FCFCFC", bg: "#99CED4" }}
-            >
+            <Tab rounded="full" fontSize={{ base: "sm", md: "md" }} _selected={{ color: "#FCFCFC", bg: "#99CED4" }}>
               プレビュー
             </Tab>
           </TabList>
           {props.pageType === "post" && (
             <Button variant="outline" w="max-content" hidden>
-              <Icon
-                as={AiFillGithub}
-                h="10"
-                w="10"
-                pr="2"
-                fontSize={{ base: "sm", md: "md" }}
-              />
+              <Icon as={AiFillGithub} h="10" w="10" pr="2" fontSize={{ base: "sm", md: "md" }} />
               GitHubから読み込む
             </Button>
           )}
         </Stack>
         <TabPanels>
           <TabPanel p="0" pt="4">
-            <MarkdownInput
-              text={props.text}
-              validText={props.validText}
-              handleText={props.handleText}
-            />
+            <MarkdownInput text={props.text} validText={props.validText} handleText={props.handleText} />
           </TabPanel>
           <TabPanel p="0" pt="4">
             <MarkdownPreview text={props.text} isFeedback={false} />
@@ -115,19 +84,9 @@ const MarkdownForm = (props: MarkdownFormProps): JSX.Element => {
       </Tabs>
       <HStack justify="space-between">
         {/* 下のinputの代わりの画像変更用ボタン */}
-        <Button
-          leftIcon={<BsImage />}
-          variant="outline"
-          onClick={onClickAddButton}
-        >
+        <Button leftIcon={<BsImage />} variant="outline" onClick={onClickAddButton}>
           {/* 上のButtonをクリックするとinputもクリックされる */}
-          <input
-            hidden
-            ref={imageInputRef}
-            type="file"
-            accept="image/*"
-            onChange={handleImage}
-          />
+          <input hidden ref={imageInputRef} type="file" accept="image/*" onChange={handleImage} />
           画像を追加
         </Button>
         <Button variant="outline" onClick={props.handlePost}>
