--- src/components/index.ts
@@ -1,25 +1,0 @@
-export { default as Product } from "./product/Product";
-export { default as LinkFavorite } from "./product/LinkFavorite";
-export { default as EditDeleteButton } from "./product/EditDeleteButton";
-export { default as Search } from "./Search";
-export { default as Post } from "./post/Post";
-export { default as User } from "./user/User";
-export { default as UserEdit } from "./userEdit/UserEdit";
-export { default as Header } from "./header/Header";
-export { default as HeaderMenu } from "./header/HeaderMenu";
-export { default as HeaderMenuUser } from "./header/HeaderMenuUser";
-export { default as GithubIcon } from "./common_part/GithubIcon";
-export { default as TwitterIcon } from "./common_part/TwitterIcon";
-export { default as ProductIcon } from "./common_part/ProductIcon";
-export { default as TagIcon } from "./common_part/TagIcon";
-export { default as MarkdownForm } from "./common_part/MarkdownForm";
-export { default as MarkdownInput } from "./common_part/MarkdownInput";
-export { default as MarkdownPreview } from "./common_part/MarkdownPreview";
-export { default as DisplayProduct } from "./common_part/DisplayProduct";
-export { default as DisplayProducts } from "./common_part/DisplayProducts";
-export { default as Favorite } from "./common_part/Favorite";
-export { default as SearchCondition } from "./Home/SearchCondition";
-export { default as postImage } from "./common_part/postImage";
-export type { default as DisplayProductProps } from "./common_part/DisplayProductProps";
-export { default as DisplayUserInfo } from "./user/DisplayUserInfo";
-export { default as DisplayUserProductList } from "./user/DisplayUserProductList";
