+++ src/components/search/Search.tsx
@@ -3,8 +3,8 @@ import { Input, InputGroup, InputRightElement, VStack, IconButton, SimpleGrid, B
 import { BrowserRouter, useHistory } from "react-router-dom";
 import { MdSearch } from "react-icons/md";
 import { QueryDocumentSnapshot, DocumentData } from "firebase/firestore";
-import { fetchAllTagList } from "../firebase/firestore";
-import { useAppDispatch } from "../hooks/hooks";
+import { fetchAllTagList } from "../../firebase/firestore";
+import { useAppDispatch } from "../../hooks/hooks";
 
 const Search = (): JSX.Element => {
   const dispatch = useAppDispatch();
