+++ src/components/userEdit/UserEdit.tsx
@@ -17,7 +17,10 @@ import { useHistory } from "react-router-dom";
 import { fetchUserInfo, postUserInfo } from "../../firebase/firestore";
 import app from "../../base";
 import { AuthContext } from "../../auth/AuthProvider";
-import { GithubIcon, TwitterIcon, postImage } from "..";
+// import { GithubIcon, TwitterIcon, postImage } from "..";
+import GithubIcon from "../common_part/GithubIcon";
+import TwitterIcon from "../common_part/TwitterIcon";
+import PostImage from "../common_part/PostImage";
 
 const storage = getStorage(app);
 
@@ -127,7 +130,7 @@ const UserEdit = (): JSX.Element => {
 
     // FileをStorageに保存し、アイコン名とURLをstateにセット
     const icon = event.target.files[0];
-    const newIconName = await postImage(icon, "user_icon", true);
+    const newIconName = await PostImage(icon, "user_icon", true);
     const newIconRef = ref(storage, newIconName);
     const downloadUrl = await getDownloadURL(newIconRef);
     setUserIconUrl(downloadUrl);
